--- eupi-shutter/src/main/java/com/euphonyio/eupi_shutter/ui/viewmodel/ShutterViewModel.kt
@@ -1,6 +1,0 @@
-package com.euphonyio.eupi_shutter.ui.viewmodel
-
-import androidx.lifecycle.ViewModel
-
-class ShutterViewModel : ViewModel(){
-}
